Update cell.rb
I got rid of some white space so the ends were all one after the other. I did also change the ship method so that the cruiser wasn't hard coded in since we will also have different ships that will be placed on the board and want it to be dynamic. Thank you and looks great! :)